Standardize some type usage to prevent confusion of devs with GCC background.

+++ avs_core/core/audio.cpp
@@ -1230,7 +1230,7 @@ void __stdcall ResampleAudio::GetAudio(void* buf, __int64 start, __int64 count, 
 	else if (offset > 0)
 	  memmove(srcbuffer, srcbuffer+offset*ch, overlap*ch<<1); // slow
 
-	last_samples= max<long long>(overlap, source_samples);                     // Samples for next time
+	last_samples= max<__int64>(overlap, source_samples);                     // Samples for next time
 
     if (source_samples-overlap > 0)                                 // Get the rest of the source samples
 	  child->GetAudio(&srcbuffer[overlap*ch], last_start+overlap, source_samples-overlap, env);
@@ -1371,7 +1371,7 @@ nofix:
 	  overlap = 0;
 	else if (offset > 0)
 	  memcpy(fsrcbuffer, fsrcbuffer+offset*ch, overlap*ch<<2);
-	last_samples= max<long long>(overlap, source_samples);
+	last_samples= max<__int64>(overlap, source_samples);
 
     if (source_samples-overlap > 0)
 	  child->GetAudio(&fsrcbuffer[overlap*ch], last_start+overlap, source_samples-overlap, env);
